# api-spec/src/main/resources/api-specification.yaml
openapi: 3.0.2
info:
  version: "1.0.0"
  title: Api-Transportes
  description: |
    Esta Api foi criada como prte integramde do trabalho de conclsão se curso da
    Faculdade Senac Florianópolis. Esta OAS3 descreve todos os endpoins disponíveis
    para utilização pelo consumidor desta aplicação
  termsOfService: "https://www.apache.org/licenses/LICENSE-2.0"
  contact:
    email: leonardoduarte1305@gmail.com
  license:
    name: Licença Apache
    url: "https://www.apache.org/licenses/LICENSE-2.0"
...
servers:
  - description: Personalizado
    url: "{protocolo}://{host}:{porta}{raiz}"
    variables:
      protocolo:
        default: http
        enum:
          - http
          - https
      host:
        default: localhost
        enum:
          - localhost
          - 127.0.0.1
      porta:
        default: '8080'
        enum:
          - '8080'
          - '8081'
          - '8082'
      raiz:
        default: /
        enum:
          - /
          - /v1
paths:
...
  /veiculos:
    post:
      operationId: criarVeiculo
      summary: Cria um veiculo
      description: "Cria um veículo com base nos dados fornecidos."
      tags:
        - Veiculos
      requestBody:
        description: "Um objeto UpsertVeiculo deve ser fornecido"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertVeiculo'
      responses:
        201:
          headers:
            Location:
              description: "Localização do veiculo criado"
              schema:
                type: number
                format: int32
                example: "/veiculos/{id}"
          description: "Created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Veiculo'
    get:
      operationId: listaTodosVeiculos
      summary: Lista todos os veiculos
      description: "Lista todos os veiculos ordenados por ID."
      tags:
        - Veiculos
      responses:
        200:
          description: "Lista dos veiculos ordenados por ID."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Veiculo'
  /veiculos/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    put:
      operationId: editarVeiculo
      summary: Edita um veiculo
      description: "Altera o cadastro de um veiculo com base no seu ID e nos dados fornecidos."
      tags:
        - Veiculos
      requestBody:
        description: "Um objeto UpsertVeiculo deve ser fornecido"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertVeiculo'
      responses:
        201:
          description: "Updated"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Veiculo'
    delete:
      operationId: excluirVeiculo
      summary: Exclui um veiculo
      description: "Exclui o registro de um veículo específico."
      tags:
        - Veiculos
      responses:
        204:
          description: "Deleted"
        400:
          description: "Id inválido"
    get:
      operationId: trazVeiculoPorId
      summary: Traz um veiculo por id
      description: "Devolve um veículo específico, selecionando-o pelo seu ID."
      tags:
        - Veiculos
      responses:
        200:
          description: "Devolve um veiculo específico buscando por ID."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Veiculo'
...
  /viagens:
    post:
      operationId: criarViagem
      summary: Cria uma viagem
      description: "Cria uma Viagem com base nos dados fornecidos."
      tags:
        - Viagens
      requestBody:
        description: "Um objeto UpsertViagem deve ser fornecido"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertViagem'
      responses:
        201:
          headers:
            Location:
              description: "Localização da viagem criada"
              schema:
                type: number
                format: int32
                example: "/viagens/{id}"
          description: "Created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viagem'
    get:
      description: "Lista todas as viagens cadastradas ordenadas por ID."
      summary: Lista todas as viagens
      operationId: listaTodasViagens
      tags:
        - Viagens
      responses:
        200:
          description: "Lista das viagens cadastradas ordenadas por ID."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Viagem'
  /viagens/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      operationId: trazViagemPorId
      summary: Traz uma viagem por id
      description: "Devolve uma viagem específica, selecionando-a pelo seu ID."
      tags:
        - Viagens
      responses:
        200:
          description: "Success."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viagem'
    put:
      operationId: editarViagem
      summary: Edita uma viagem
      description: "Altera o cadastro de uma viagem com base no seu ID e nos dados fornecidos."
      tags:
        - Viagens
      requestBody:
        description: "Um objeto UpsertViagem deve ser fornecido"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertViagem'
      responses:
        201:
          description: "Updated"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viagem'
    delete:
      operationId: excluirViagem
      summary: Exclui uma viagem
      description: "Exclui o registro de uma viagem específica.
              Aremoção é apenas lógica, o registro continua persistido no banco de dados
              e pode ser consultado apenas utilizando queries contra o banco."
      tags:
        - Viagens
      responses:
        204:
          description: "Deleted"
        400:
          description: "Id inválido"
  /viagens/{id}/destinos:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      operationId: trazDestinosDaViagem
      summary: Lista todos os destinos de uma viagem
      description: "Devolve a lista de destinos desta viagem específica."
      tags:
        - Viagens
      responses:
        200:
          description: "Devolve a lista de destinos desta viagem específica."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Destino'
  /viagens/{id}/confirmacao:
    parameters:
      - $ref: '#/components/parameters/id'
    post:
      operationId: confirmaViagem
      summary: Confirma uma viagem
      description: "Confirma uma viagem e todos os destinos cadastrados até agora."
      tags:
        - Viagens
      requestBody:
        description: "Um objeto Confirmacao deve ser fornecido"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Confirmacao'
      responses:
        204:
          description: "Viagem foi confirmada."
  /viagens/{id}/encerramento:
    parameters:
      - $ref: '#/components/parameters/id'
    post:
      operationId: encerraViagem
      summary: Encerra uma viagem
      description: "Confirma uma viagem e todos os destinos cadastrados até agora."
      tags:
        - Viagens
      requestBody:
        description: "Um objeto Encerramento deve ser fornecido"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Encerramento'
      responses:
        204:
          description: "Viagem foi encerrada."
  /viagens/{id}/exportar:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      operationId: relatorioDeViagem
      summary: Download do relatrório de uma viagem
      description: "Devolve um relatório sobre esta viagem específica."
      tags:
        - Viagens
      responses:
        200:
          description: "Conteúdo anexado."
          content:
            application/pdf:
              schema:
                type: object
...
  /usuarios:
    post:
      operationId: criarUsuario
      summary: Cria um usuário
      description: "Cria um usuário no Keycloak que consegue acessar os recursos da aplicação de acordo
        com sua Role"
      tags:
        - Usuarios
      requestBody:
        description: "Um objeto UpsertUsuario deve ser fornecido"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertUsuario'
      responses:
        201:
          description: "Localização do usuario criado"
  /usuarios/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: "O id da entidade que desejamos analisar"
        schema:
          type: string
    delete:
      operationId: excluirUsuario
      summary: Exclui um usuário
      description: "Exclui o registro de um usuario específico."
      tags:
        - Usuarios
      responses:
        204:
          description: "Excluido"
        400:
          description: "Id inválido"
components:
  schemas:
...
    UpsertUsuario:
      type: object
      properties:
        username:
          type: string
          example: "talita01"
        password:
          type: string
          example: "abc123"
          format: password
          writeOnly: true
        nome:
          type: string
          example: "Talita"
        sobrenome:
          type: string
          example: "da Silva"
        email:
          type: string
          example: "meuemail@gmail.com"
          format: email
        role:
          type: string
          enum:
            - ADMIN
            - USER
          example: "ADMIN"
      required:
        - email
        - roles
        - username
        - password
...
    AtributosVeiculo:
      type: object
      properties:
        modelo:
          type: string
          description: "Modelo do veículo."
          example: "Hyundai HR Baú"
        marca:
          type: string
          description: "Marca da montadora do veículo."
          example: "Hyundai"
        placa:
          type: string
          description: "Placa do veículo, atributo único, padrão Mercosul ou padrão brasileiro anterior."
          pattern: '[A-Z]{3}-\d{4}|[A-Z]{3}\d[A-Z]\d{2}'
        ano:
          type: integer
          description: "Ano de fabricação do veículo."
          example: "2020"
        renavan:
          type: number
          description: "Número de identificação do veículo."
          example: "3214569871"
          pattern: '\d{10}'
        tamanho:
          type: string
          description: "Tamanho do veículo."
          example: "3,58m"
      required:
        - modelo
        - marca
        - placa
        - ano
        - renavan
        - tamanho
    UpsertVeiculo:
      allOf:
        - $ref: '#/components/schemas/AtributosVeiculo'
    Veiculo:
      type: object
      allOf:
        - $ref: '#/components/schemas/UpsertVeiculo'
      properties:
        id:
          type: integer
          description: "Atributo que identifica cada veículo de maneira única."
          example: "102"
...
    MaterialQuantidadeSetor:
      type: object
      properties:
        materialId:
          type: integer
          description: "Id do material que deve ser transportado."
        quantidade:
          type: integer
          description: "Quantidade do material que deve ser transportado."
        setorDestino:
          type: integer
          description: "Setor onde o materia deve ser entregue na unidade de destino."
    UpsertDestino:
      type: object
      description: "Sede por onde o transporte deve passar e os materiais que ele deve deixar nesta parada."
      properties:
        sedeId:
          type: integer
          description: "Id da sede para onde devem ir os materiais transportados"
        materiaisQntdSetor:
          type: array
          items:
            $ref: '#/components/schemas/MaterialQuantidadeSetor'
            # O status será adicionado automaticamente no backend
            # -> se o ADMIN criar este destino ele terá status = true
            # -> se outro usuário criar este destino ele terá status = false até que o ADMIN confirme
    Destino:
      type: object
      allOf:
        - $ref: '#/components/schemas/UpsertDestino'
      properties:
        id:
          type: integer
          description: "Atributo que identifica cada destino de maneira única."
          example: "3"
        status:
          $ref: '#/components/schemas/Confirmacao'
    Confirmacao:
      type: object
      properties:
        confirmacao:
          type: string
          description: "Se passado por um ADMIN confirmará ou não a viagem, caso contrário, nada acontece."
          example: "CONFIRMADO"
          enum:
            - CONFIRMADO
            - NAO_CONFIRMADO
    Encerramento:
      type: object
      properties:
        encerrado:
          type: string
          description: "Se passado por um ADMIN encerrará a viagem, caso contrário, nada acontece."
          example: "ENCERRAR"
          enum:
            - ENCERRAR
    UpsertViagem:
      type: object
      description: "Todas as informacoes necessarias para o agendamento de uma viagem"
      required:
        - datetimeSaida
      properties:
        motoristaId:
          type: integer
          description: "Id único que identifica o motorista desta viagem."
        veiculoId:
          type: integer
          description: "Id único que identifica o veiculo usado nesta viagem."
        destinos:
          type: array
          items:
            type: integer
        datetimeSaida:
          type: string
          description: "Data e hora de partida da viagem."
        datetimeVolta:
          type: string
          description: "Data e hora de partida da viagem."
        sede:
          type: integer
          description: "Atributo que identifica a sede de onde parte a viagem."
          example: "32"
    Viagem:
      type: object
      allOf:
        - $ref: '#/components/schemas/UpsertViagem'
      properties:
        status:
          $ref: '#/components/schemas/Confirmacao'
        encerrado:
          type: string
          description: "Atributo que identifica se a viagem já foi encerrada possibilidades: ENCERRADO ou NAO_ENCERRADO"
          example: "ENCERRADO"
        sede:
          type: integer
          description: "Atributo que identifica a sede de onde parte a viagem."
          example: "32"
        id:
          type: integer
          description: "Atributo que identifica cada viagem de maneira única."
          example: "44"
...
  parameters:
    id:
      name: id
      in: path
      required: true
      description: "O id da entidade que desejamos analisar"
      schema:
        type: integer
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
#    BasicAuth:
#      type: http
#      scheme: basic
security:
  - BearerAuth: [ ]
#  - BasicAuth: [ ]